Conflicts:
	Raven.Database/Queries/SuggestionQueryIndexExtension.cs
	Raven.Studio/Models/EditableDocumentModel.cs

+++ Raven.Tests/RavenTest.cs
@@ -33,7 +33,7 @@ namespace Raven.Tests
 
 		private string path;
 
-		public EmbeddableDocumentStore NewDocumentStore(string storageType = "munin", bool inMemory = true, int? allocatedMemory = null)
+		public EmbeddableDocumentStore NewDocumentStore(string storageType = "munin", bool inMemory = true, int? allocatedMemory = null, bool deleteExisting = true)
 		{
 			path = Path.GetDirectoryName(Assembly.GetAssembly(typeof(DocumentStoreServerTests)).CodeBase);
 			path = Path.Combine(path, "TestDb").Substring(6);
@@ -52,7 +52,7 @@ namespace Raven.Tests
 			ModifyStore(documentStore);
 			ModifyConfiguration(documentStore.Configuration);
 
-			if (documentStore.Configuration.RunInMemory == false)
+			if (documentStore.Configuration.RunInMemory == false && deleteExisting)
 				IOExtensions.DeleteDirectory(path);
 			documentStore.Initialize();
